refactor(data): centralize business domain data
